refactor: legacy 패키지를 제거하고 tobe 패키지를 도메인 패키지로 이동
