+++ src/test/java/bowling/domain/NormalFrameTest.java
@@ -3,75 +3,53 @@ package bowling.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.Arrays;
+import static org.assertj.core.api.Assertions.assertThat;
 
-import static org.assertj.core.api.Assertions.*;
+@DisplayName("NormalFrame: 1~9 프레임 테스트")
+public class NormalFrameTest {
 
-@DisplayName("NormalFrame 로직 테스트")
-class NormalFrameTest {
-
+    @DisplayName("1번 던져서 넘긴 핀이 10개면 Finish 반환")
     @Test
-    public void 스트라이크() {
-        Frame frame = new NormalFrame();
-        frame.bowling(10);
-        assertThat(frame.getResult()).isEqualTo(Arrays.asList(Shot.STRIKE));
+    public void bowling_Strike_ReturnFinish() {
+        Frame frame = NormalFrame.first();
+        assertThat(frame.bowling(new Pin(10))).isEqualTo(State.Finish);
     }
 
+    @DisplayName("1번 던져서 넘긴 핀이 10개 미만 NotFinish 반환")
     @Test
-    public void 스페어() {
-        Frame frame = new NormalFrame();
-        frame.bowling(5);
-        frame.bowling(5);
-        assertThat(frame.getResult()).isEqualTo(Arrays.asList(Shot.FIVE, Shot.SPARE));
+    public void bowling_Miss_ReturnNotFinish() {
+        Frame frame = NormalFrame.first();
+        assertThat(frame.bowling(new Pin(5))).isEqualTo(State.NotFinish);
     }
 
+    @DisplayName("2번 던지면 Finish 반환")
     @Test
-    public void 미스() {
-        Frame frame = new NormalFrame();
-        frame.bowling(5);
-        frame.bowling(3);
-        assertThat(frame.getResult()).isEqualTo(Arrays.asList(Shot.FIVE, Shot.THREE));
+    public void bowling_TwoTime_ReturnFinish() {
+        Frame frame = NormalFrame.first();
+        frame.bowling(new Pin(3));
+        assertThat(frame.bowling(new Pin(3))).isEqualTo(State.Finish);
     }
 
+    @DisplayName("라운드는 1부터 시작")
     @Test
-    public void 거터() {
-        Frame frame = new NormalFrame();
-        frame.bowling(0);
-        frame.bowling(0);
-        assertThat(frame.getResult()).isEqualTo(Arrays.asList(Shot.GUTTER, Shot.GUTTER));
+    public void 라운드는_1부터_시작() {
+        Frames frames = new Frames(new Player("jae"));
+        assertThat(frames.getCurrentFrameNo()).isEqualTo(1);
     }
 
+    @DisplayName("프레임이 종료되면 프레임이 증가한다")
     @Test
-    public void 핀카운트가_10개를_넘음() {
-        Frame frame = new NormalFrame();
-        assertThatThrownBy(() ->
-                frame.bowling(11)
-        ).isInstanceOf(IllegalArgumentException.class);
+    public void 프레임이_종료되면_프레임이_증가한다() {
+        Frames frames = new Frames(new Player("jae"));
+        frames.bowling(new Pin(10));
+        assertThat(frames.getCurrentFrameNo()).isEqualTo(2);
     }
 
+    @DisplayName("프레임이 종료되지않으면 프레임이 유지된다")
     @Test
-    public void 스트라이크라면_프레임종료() {
-        Frame frame = new NormalFrame();
-        assertThat(frame.bowling(10)).isTrue();
-    }
-
-    @Test
-    public void 스페어라면_프레임종료() {
-        Frame frame = new NormalFrame();
-        frame.bowling(5);
-        assertThat(frame.bowling(5)).isTrue();
-    }
-
-    @Test
-    public void 스페어처리해야되면_프레임종료안됨() {
-        Frame frame = new NormalFrame();
-        assertThat(frame.bowling(5)).isFalse();
-    }
-
-    @Test
-    public void 미스라면_프레임종료() {
-        Frame frame = new NormalFrame();
-        frame.bowling(5);
-        assertThat(frame.bowling(3)).isTrue();
+    public void 프레임이_종료되지않으면_프레임이_유지된다() {
+        Frames frames = new Frames(new Player("jae"));
+        frames.bowling(new Pin(5));
+        assertThat(frames.getCurrentFrameNo()).isEqualTo(1);
     }
 }
